Improve Pull action for Paparazzi
